pyproject.toml
[tool.poetry]
name = "dashboardapi"
version = "0.1.0"
description = ""
authors = ["M. Hatem"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.6"
drf-yasg = "^1.21.7"
django-rest-swagger = "^2.2.0"
django-model-utils = "^4.3.1"
django-phonenumber-field = "^7.2.0"
phonenumbers = "^8.13.23"
markdown = "^3.5"
djangorestframework-simplejwt = "^5.3.0"
python-dotenv = "^1.0.0"
django-environ = "^0.11.2"
django-cors-headers = "^4.3.0"

[tool.poetry.group.dev.dependencies]
django-rest-framework = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
